link_bio/requirements.txt
pip==24.0
reflex==0.4.4
python-dotenv==1.0.1